Add error hook type
