+++ openfeature/sdk/src/main/scala/io/cardell/openfeature/Hooks.scala
@@ -69,33 +69,6 @@ case class HookContext(
 
 }
 
-sealed trait Hook[F[_]]
-
-trait BeforeHook[F[_]] extends Hook[F] {
-
-  def apply(
-      context: HookContext,
-      hints: HookHints
-  ): F[Option[EvaluationContext]]
-
-}
-
-object BeforeHook {
-
-  def apply[F[_]](
-      f: (HookContext, HookHints) => F[Option[EvaluationContext]]
-  ): BeforeHook[F] =
-    new BeforeHook[F] {
-
-      def apply(
-          context: HookContext,
-          hints: HookHints
-      ): F[Option[EvaluationContext]] = f(context, hints)
-
-    }
-
-}
-
 object HookHints {
   def empty: HookHints = Map.empty
 }
@@ -124,5 +97,13 @@ object Hooks {
 
     aux(hooks, context).map(_.getOrElse(context.evaluationContext))
   }
+
+  def runErrors[F[_]: Monad](
+      hooks: List[ErrorHook[F]]
+  )(
+      context: HookContext,
+      hints: HookHints,
+      error: Throwable
+  ): F[Unit] = hooks.traverse(_.apply(context, hints, error)).void
 
 }
